funcfolder/requirements.txt
azure-functions
azure-storage-blob
fpdf
certifi
fastapi
uvicorn
python-dotenv
requests
azure-servicebus
certifi
passlib[bcrypt]
python-jose
cryptography
gunicorn
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
SQLAlchemy==2.0.25
pymysql==1.1.0
